Hopefully fixes pipeline, moved file structure
